classifier/serializers.py
from rest_framework import serializers

from classifier.models import Image


class ImageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Image
        fields = '__all__'
        read_only_fields = ('classification',)
